+++ seed/static/seed/js/seed.js
@@ -77,6 +77,7 @@ angular.module('BE.seed.controllers', [
   'BE.seed.controller.pairing',
   'BE.seed.controller.pairing_settings',
   'BE.seed.controller.profile',
+  'BE.seed.controller.rename_column_modal',
   'BE.seed.controller.security',
   'BE.seed.controller.settings_profile_modal',
   'BE.seed.controller.show_populated_columns_modal',