+++ CHANGELOG.md
@@ -1,7 +1,16 @@
-# SEED Version 2.6.0 - Beta
+# SEED Version 2.6.0
 
 2.6.0-beta includes support for meters and time series data storage. In order to use this release, you must first install [timescaledb](https://docs.timescale.com/v1.2/getting-started). Note that this is a Beta release and should not be used on production datasets until we have completed testing.
 
+# SEED Version 2.5.2
+
+Date Range: 04/15/19 - 05/04/19
+
+Closed Issues and Features:
+- Feature [#1855]( https://github.com/SEED-platform/seed/issues/1855 ), Add ULID as a matching field on taxlots
+- Feature [#1856]( https://github.com/SEED-platform/seed/issues/1856 ), Import building and taxlot footprints from WKT format
+- Feature [#1861]( https://github.com/SEED-platform/seed/issues/1861 ), Allow renaming of columns
+- Fixed [#1867]( https://github.com/SEED-platform/seed/issues/1867 ), Allow Bldg to be parsed as part of the normalized address
 
 # SEED Version 2.5.1